+++ invisibleroads_users/settings.py
@@ -1,0 +1,4 @@
+from invisibleroads_posts.libraries.configuration import Settings
+
+
+S = Settings()